refactor: Move message box components to separate files
